third_party/WebKit/LayoutTests/fast/dom/shadow/multiple-shadowroot-adopt-expected.txt

CONSOLE WARNING: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
This test checks multiple shadow trees can be moved into another document.

PASS document is sr1.ownerDocument
PASS sr1.ownerDocument is sr2.ownerDocument
PASS sr1.ownerDocument is iframeDoc
PASS sr2.ownerDocument is iframeDoc
PASS successfullyParsed is true

TEST COMPLETE
